+++ src/main/java/io/github/mechtasnezhevna/createpatina/mixin/recipe/SandPaperItemMixin.java
@@ -3,6 +3,7 @@ package io.github.mechtasnezhevna.createpatina.mixin.recipe;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.simibubi.create.content.equipment.sandPaper.SandPaperItem;
+import io.github.mechtasnezhevna.createpatina.PatinaConfig;
 import io.github.mechtasnezhevna.createpatina.block.PatinaBlock;
 import io.github.mechtasnezhevna.createpatina.util.OxidizeUtil;
 import net.minecraft.core.BlockPos;
@@ -27,7 +28,8 @@ public abstract class SandPaperItemMixin {
         BlockState oldState = level.getBlockState(pos);
         if (oldState.getBlock() instanceof PatinaBlock
                 || newState.getBlock() instanceof PatinaBlock) {
-            OxidizeUtil.replaceWithState(oldState, newState, level, pos);
+            OxidizeUtil.applyToolWeathering(oldState, newState, level, pos,
+                    PatinaConfig.CONFIG.WEATHER_WHOLE_FLUID_TANK_WITH_TOOLS.get());
             return true;
         }
         return original.call(level, pos, newState);